+++ ComputationalMethods/sources/Vector.cpp
@@ -29,7 +29,7 @@ size_t Vector::Size() const
 Vector operator+(const Vector a, const Vector b)
 {
 	Vector c(a.Size());
-	for (int i = 0; i < a.Size(); ++i)
+	for (size_t i = 0; i < a.Size(); ++i)
 	{
 		c[i] = a[i] + b[i];
 	}
@@ -60,7 +60,7 @@ void Vector::Swap(const int i, const int j)
 
 std::ostream& operator<<(std::ostream& os, const Vector& a)
 {
-	for(int i = 0; i < a.Size(); ++i)
+	for(size_t i = 0; i < a.Size(); ++i)
 	{
 		os<<a[i]<<'\n';
 	}
@@ -72,7 +72,7 @@ std::istream& operator>>(std::istream& is, Vector& a)
 	size_t size;
 	is>>size;
 	a = Vector(size);
-	for(int i = 0; i < a.Size(); ++i)
+	for(size_t i = 0; i < a.Size(); ++i)
 		is>>a[i];
 	return is;
 }
@@ -80,7 +80,7 @@ std::istream& operator>>(std::istream& is, Vector& a)
 double operator*(const Vector& a, const Vector& b)
 {
 	double s = 0.0;
-	for(int i = 0; i < a.Size(); ++i)
+	for(size_t i = 0; i < a.Size(); ++i)
 		s += a[i] * b[i];
 	return s;
 }
@@ -88,7 +88,7 @@ double operator*(const Vector& a, const Vector& b)
 Vector operator-(const Vector& a, const Vector& b)
 {
 	Vector c(a);
-	for(int i = 0; i < a.Size(); ++i)
+	for(size_t i = 0; i < a.Size(); ++i)
 		c[i] -= b[i];
 	return c;
 }
@@ -96,14 +96,14 @@ Vector operator-(const Vector& a, const Vector& b)
 double Vector::Norm() const
 {
 	double n = 0.0;
-	for(int i = 0; i < vector.size(); ++i)
+	for(size_t i = 0; i < vector.size(); ++i)
 		n += vector[i] * vector[i];
 	return sqrt(n);
 }
 
 Vector operator*(const double a, Vector& b)
 {
-	for(int i = 0; i < b.Size(); ++i)
+	for(size_t i = 0; i < b.Size(); ++i)
 		b[i] = a * b[i];
 	return b;
 }
